Update resource tweaking for 0.24.
UIPartActionResourceEditorImproved.OnSliderChanged seems to be a copy of
the stock code with a couple of enhancements. The stock code now sends a
GameEvent.onEditorShipModified event (and updates the displayed value).

This fixes the breakage of the vessel cost widget when adjusting the amount
of resource in a part in the editor.