Create tagged versions of component classes.
It seems that Unity's AddComponent<> doesn't actually use the type any
more, but rather gets the type name then looks up the type. This rather
breaks KAE's multi-version class hunting as it's difficult to predict which
assembly will be used for the class.

It seems that different assembly names might help, but that makes things
difficult for KAE, so instead create classes derived from the working
classes with names including the assembly version. It is these derived
classes that will be used for AddComponent and KSPAddon.